Introduce the `metricscollectiontime` meta metric
The `metricscollectiontime` indicates the system time (in millis) when
the last metrics collection was performed. The git repo metrics plugin,
in order to spare system's resources, typically collects metrics
periodically and then seves them from cache. The `metricscollectiontime`
metric tracks the collection time so that it can be easily determined
if/when metrics were collected.

Bug: Issue 427093248
Change-Id: I4b22920bfcdaeb55259842261a0cf8cab7601aa8